# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for managing transaction.\n\u003cbr/\u003e\u003cb\u003eDeveloper:\u003c/b\u003e Muhammad Brilian Satria Utama\n\u003cb\u003eEnvironment:\u003c/b\u003e Development",
        "title": "E Wallet API (Transaction Service)",
        "contact": {},
        "version": "0.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/transaction/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Processes a new transaction, updating the user's current balance and generating a transaction record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Create Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payload create transaction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ewallet-transaction_internal_dto_transaction_dto.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ewallet-transaction_internal_dto_transaction_dto.CreateTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.Code": {
            "type": "string",
            "enum": [
                "UNKNOWN_ERROR",
                "INTERNAL_SERVER_ERROR",
                "BAD_REQUEST",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "NOT_FOUND",
                "USER_NOT_FOUND",
                "SESSION_NOT_FOUND",
                "INVALID_PASSWORD",
                "WALLET_NOT_FOUND",
                "INVALID_STATUS_TRANSITION",
                "DUPLICATE_REFERENCE",
                "TRANSACTION_NOT_FOUND",
                "INSUFFICIENT_BALANCE"
            ],
            "x-enum-varnames": [
                "ErrCodeUnknownError",
                "ErrCodeInternalServerError",
                "ErrCodeBadRequest",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeNotFound",
                "ErrCodeUserNotFound",
                "ErrCodeSessionNotFound",
                "ErrCodeInvalidPassword",
                "ErrCodeWalletNotFound",
                "ErrCodeInvalidStatusTransition",
                "ErrCodeDuplicateReference",
                "ErrCodeTransactionNotFound",
                "ErrCodeInsufficientBalance"
            ]
        },
        "ewallet-transaction_internal_domain_transaction.Type": {
            "type": "string",
            "enum": [
                "TOPUP",
                "PURCHASE",
                "REFUND"
            ],
            "x-enum-varnames": [
                "Topup",
                "Purchase",
                "Refund"
            ]
        },
        "ewallet-transaction_internal_dto_transaction_dto.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "description",
                "transaction_type"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 10000
                },
                "description": {
                    "type": "string",
                    "example": "Top up to buy internet package"
                },
                "transaction_type": {
                    "enum": [
                        "TOPUP",
                        "PURCHASE",
                        "REFUND"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/ewallet-transaction_internal_domain_transaction.Type"
                        }
                    ],
                    "example": "TOPUP"
                }
            }
        },
        "ewallet-transaction_internal_dto_transaction_dto.CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "reference": {
                    "type": "string"
                },
                "transaction_status": {
                    "type": "string"
                }
            }
        },
        "response.BadRequestResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "$ref": "#/definitions/errors.Code"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ValidationErrorField"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "$ref": "#/definitions/errors.Code"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {}
            }
        },
        "response.ValidationErrorField": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}